+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-06-28
+### Changed
+- parameters are stored using ID instead of full query
+
+### Added
+- input and output groups in data model
+
 ## [1.1.0] - 2024-06-06
 ### Changed
 - searchable multi select dropdowns for dashboard filters